move added test to only build for amd64
