+++ go/protocol/keybase1/extras_test.go
@@ -7,8 +7,6 @@ import (
 	"errors"
 	"testing"
 	"time"
-
-	"github.com/stretchr/testify/require"
 )
 
 func TestTime(t *testing.T) {
@@ -134,12 +132,4 @@ func TestTeamNameFromString(t *testing.T) {
 			}
 		}
 	}
-}
-
-func TestSize(t *testing.T) {
-	require.Equal(t, UID("a").Size(), 9)
-	require.Equal(t, TeamID("a").Size(), 9)
-	require.Equal(t, UserOrTeamID("a").Size(), 9)
-	kid := KID("a")
-	require.Equal(t, kid.Size(), 9)
 }
